requirements.txt
streamlit
together
requests
speechrecognition
pillow
